# docs/swagger.yaml
swagger: '2.0'
info:
  title: Gryphon API
  description: Store and load case management fragments using BPMN and a JSON API.
  version: 1.0.0
schemes:
  - http
basePath: /api
produces:
  - application/json
paths:
...
  /domainmodel/{id}:
    get:
      summary: Returns a domain model based on its ID.
      description: |
        Returns a full domain model including all attributes.
      parameters:
        - name: id
          in: path
          description: The domain models id.
          type: number
          format: integer
          required: true
      tags:
        - DomainModel
      responses:
        '200':
          description: The domain model if found.
          schema:
            $ref: '#/definitions/DomainModel'
        '404':
          description: Model not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Updates a domain model.
      description: |
        This endpoint updates a domain model based on its ID. If the attributes differ from earlier version, the revision will be increased. Also, all scenarios that reference a domain model will get increased an revision. The repository doesn't care if the id given in the object differs from the id given in the path. It will always use the paths ID.
      parameters:
        - name: id
          in: path
          description: The domain models id.
          required: true
          type: number
          format: integer
        - name: domainmodel
          in: body
          description: The domainmodel-content.
          required: true
          schema:
            $ref: '#/definitions/DomainModel'
      tags:
        - Fragment
      responses:
        '200':
          description: The domain model if found and update was succesfull.
          schema:
            $ref: '#/definitions/DomainModel'
        '404':
          description: Model not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /domainmodel:
    get:
      summary: |
        Returns a list of all available domain models.
        des  Returns a list of all domain models. This list might be filtered by searching
        using a name.
      parameters:
        - name: query
          in: query
          description: Might be used to filter the domain models by name.
          required: false
          type: string
          format: string
      tags:
      - DomainModel
      responses:
        '200':
          description: The list of domain models.
          schema:
            $ref: '#/definitions/DomainModelList'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Creates a new domain model
      description: |
        This endpoint stores a new domain model on the server.
      parameters:
        - name: domainmodel
          in: body
          description: The model object.
          required: true
          schema:
            $ref: '#/definitions/DomainModel'
      tags:
        - DomainModel
      responses:
        '200':
          description: The updated model, with updated id.
          schema:
            $ref: '#/definitions/DomainModel'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /scenario/{id}/export:
    get:
      summary: Exports a scenario.
      description: |
        This endpoint exports a scenario to the url given in the body object. Further additions to the body object might be possible.
        This endpoint exports the whole scenario (including fragments and domain model) in JSON format.
        The target server should respond with a 200 if everything went fine. The content of the response doesn't matter.
        The target server should expect a POST-call according to RESTful APIs.
      parameters:
        - name: id
          in: path
          description: The scenarios id.
          required: true
          type: number
          format: int32
        - name: exportID
          in: body
          description: The model object.
          required: true
          schema:
            $ref: '#/definitions/ExportRequest'
      tags:
        - Scenario
      responses:
        '200':
          description: The data that was returned by the target-server.
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Scenario doesn't exist
        '400':
          description: Target server is unreachable or doesn't respond as expected.
  /export:
    get:
      summary: |
        Returns a list of all available export-targets.
      tags:
        - Export
      responses:
        '200':
          description: The list of domain models.
          schema:
            $ref: '#/definitions/DomainModelList'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Creates a new export-target.
      description: |
        This endpoint creates a new export-target.
      parameters:
        - name: export
          in: body
          description: The export object.
          required: true
          schema:
            $ref: '#/definitions/Export'
      tags:
        - Export
      responses:
        '200':
          description: The created export-target, with updated id.
          schema:
            $ref: '#/definitions/Export'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /export/{id}:
    delete:
      summary: |
        Deletes the export-target with the given ID.
      parameters:
        - name: id
          in: path
          description: The target id.
          required: true
          type: number
          format: integer
      tags:
        - Export
      responses:
        '200':
          description: The deleted export-target.
          schema:
            $ref: '#/definitions/Export'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Creates a new export-target.
      description: |
        This endpoint updates the given export-target.
      parameters:
        - name: id
          in: path
          description: The target id.
          required: true
          type: number
          format: integer
        - name: export
          in: body
          description: The export object.
          required: true
          schema:
            $ref: '#/definitions/Export'
      tags:
        - Export
      responses:
        '200':
          description: The updated export-target.
          schema:
            $ref: '#/definitions/Export'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Export:
    type: object
    properties:
      _id:
        type: number
        format: integer
        description: 'The unique identifier of this export-target.'
      name:
        type: string
        description: 'The name of the export-target.'
      url:
        type: string
        description: 'The url of this export-target.'
...
  DomainModel:
    type: object
    required:
      - name
      - dataclasses
    properties:
      name:
        type: string
        description: 'The name of the domain model.'
      _id:
        type: number
        format: long
        description: 'The unique identifier of the domain model.'
      revision:
        type: number
        format: integer
        description: 'The domain models revision. Gets increased everytime the domain model is updated.'
      dataclasses:
        type: array
        description: 'The data classes belonging to the domain model.'
        items:
          $ref: '#/definitions/DataClass'
  DataClass:
    type: object
    properties:
      name:
        type: string
        description: 'The name of the data class.'
      _id:
        type: number
        format: long
        description: 'The unique identifier the data class.'
      is_root:
        type: boolean
        description: 'Describes whether the data class is the root element of the domain model.'
      is_event:
        type: boolean
        description: 'Describes whether the data class represents event data.'
      attributes:
        type: array
        description: 'The attributes of the data class.'
        items:
          $ref: '#/definitions/DataClassAttribute'
  DataClassAttribute:
    type: object
    properties:
      name:
        type: string
        description: 'The name of the attribute.'
      datatype:
        type: string
        description: 'The datatype of the attribute.'
...
  DomainModelList:
    type: array
    description: 'All models found.'
    items:
      $ref: '#/definitions/DomainModel'
  Error:
    type: object
  ExportRequest:
    type: object
    properties:
      exportID:
        type: number
        format: int32
        description: The url the post-call should use.